requirements.txt
cycler==0.10.0
matplotlib==1.5.2
numexpr==2.6.1
numpy==1.11.1
pandas==0.18.1
pyaml==15.8.2
pyparsing==2.1.8
python-dateutil==2.5.3
pytz==2016.6.1
PyYAML==3.11
requests==2.11.1
six==1.10.0
spotipy==2.3.8
SQLAlchemy==1.0.14
tables==3.2.3
wheel==0.26.0
